Struct vulkan_layer::test_utils::TestGlobalBuilder

pub struct TestGlobalBuilder<T: TestLayerTag = Tag<0>> { /* private fields */ }

Available on crate feature _test only.

The builder for TestGlobal.

Implementations

impl<T: TestLayerTag> TestGlobalBuilder<T>

pub const fn set_layer_mock_builder( self, layer_mock_builder: fn() -> MockTestLayer<T>, ) -> Self

Specify how MockTestLayer should be created.

Can be used to customize the behavior of pre-initialized methods, like InstanceInfo::hooked_commands.

Pre-initialized methods are almost always called, so it is usually needed to call MockTestLayer::set_default_expections right before return to provide a meaningful default behavior.

pub const fn set_layer_global_builder( self, layer_global_builder: fn() -> Global<TestLayer<T>>, ) -> Self

Specify how Global<TestLayer<T>> should be created.

Can be used to customize the behavior and expections of layer behaviors after Global is initialized, especially the global commands, e.g. GlobalHooks::create_instance.

pub const fn build(&self) -> TestGlobal<T>

Construct the TestGlobal<T>.

This is a const function, and can be directly used to initialize a static variable.