Module vulkan_layer::unstable_api

Available on crate feature unstable only.

Vulkan layer utilities that are used in the integration tests to implement mock layers. The utilities may benefit the layer implementation, but there is no stability guarantee on those APIs.

Structs

ApiVersion

Vulkan API version number.

LazyCollection

A collection wrapper, that guarantees that an empty collection can be trivially destructed.

Enums

Feature

A union type of extensions and core API version.

Traits

CheckEmpty

The CheckEmpty trait allows to check if a collection is empty.

IsCommandEnabled

Used to extend an iterator of Feature.